split off qt dev dependencies to separate group, and bump version nr
