+++ observ/__init__.py
@@ -1,4 +1,4 @@
-__version__ = "0.13.1"
+__version__ = "0.14.0"
 
 
 from .init import init, loop_factory
